fix(language-service): transform markdown links in completion items
This resolves completion item descriptions the same way as hover
descriptions are resolved.

The `transformDocumentLinkTarget` and `transformMarkdown` functions were
both moved into `utils/transform`, but their implementation remains the
same.

The fix was verified to work in MDX analyzer.

Refs https://github.com/mdx-js/mdx-analyzer/issues/394
